all routes sorted, final commit for day

+++ db/scripts/createTableConfirmation.js
@@ -4,8 +4,8 @@ async function createTableConfirmation() {
   const res = await query(`CREATE TABLE IF NOT EXISTS tree_confirmation(
       tree_id SERIAL PRIMARY KEY,
       species TEXT,
-      longitude INTEGER,
-      latitude INTEGER,
+      longitude DECIMAL(6,4),
+      latitude DECIMAL(5,4),
       planted BOOLEAN,
       date_planted DATE
       )
